fix: haproxy #369 perf issue
